Set or get options for Pupilla's preprocessing parameters

Set or get options for Pupilla's preprocessing parameters

pp_options(...)

Arguments

thresh

speed_clean Threshold (z point) above which values are marked as NA.

extend_by

speed_clean Number of samples starting from the deviant speed values that are stripped (e.g., the signal in proximity of blinks may be biased as well).

island_size

speed_clean Islands of signal in the midst of NAs are removed if smaller or equal to this threshold (amount of samples).

extend_blink

interpolate NAs are extended by these many samples prior to interpolation. This gets rid of signal that may be compromised in the close proximity of a blink.

overall_thresh

interpolate Overall quality threshold: e.g., the total amount of data that is allowed to be missing from the original vector.

consecutive_thresh

interpolate Consecutive gaps in the signal: e.g., the total amount of data that is allowed to be missing from the original vector consecutively.

spar

smooth Smoothing factor as in 'smooth.spline()'.