A convenience function to preprocess pupillometry data

The function calls, in order, 'speed_clean', 'interpolate', and 'smooth_vector' from this packages. Parameters can only be changed in the package options, i.e. 'Pupilla::pp_options()'. Warning, the best preprocessing parameters may deviate from the defaults used here, be mindful!

Usage

pre_process(vector, time)

Arguments

vector

A vector variable to be cleaned

time

A vector variable indicating the elapsed time, needed to compute velocity.

Value

A numeric vector processed as requested or as per the default. .